agent-hiring-mvp/agenthub-sdk/templates/acp_rag/requirements.txt
llama-index
acp-sdk
python-dotenv
llama-index-readers-web
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-llms-langchain
llama-index-embeddings-langchain
llama-index-llms-ollama
langchain-openai
langchain_community